requirements.txt
hyperliquid-python-sdk==0.15.0
binance-connector==3.12.0
python-binance==1.0.29
matplotlib==3.10.3
mplfinance>=0.12.10
numpy>=1.26.4